+++ compiler-core/src/parse.rs
@@ -1204,12 +1204,15 @@ where
 
                 let mut elements_after_tail = None;
                 let mut dot_dot_location = None;
-                let tail = if let Some((start, Token::DotDot, end)) = self.tok0 {
-                    dot_dot_location = Some((start, end));
-                    if !elements_end_with_comma {
+                let tail = if let Some((dot_dot_start, Token::DotDot, dot_dot_end)) = self.tok0 {
+                    dot_dot_location = Some((dot_dot_start, dot_dot_end));
+                    if !elements.is_empty() && !elements_end_with_comma {
                         self.warnings
                             .push(DeprecatedSyntaxWarning::DeprecatedListPattern {
-                                location: SrcSpan { start, end },
+                                location: SrcSpan {
+                                    start: dot_dot_start,
+                                    end: dot_dot_end,
+                                },
                             });
                     }
 
@@ -1271,6 +1274,13 @@ where
                     // No tail specified
                     None => None,
                 };
+
+                if elements.is_empty() && tail.as_ref().is_some_and(|p| p.is_discard()) {
+                    self.warnings
+                        .push(DeprecatedSyntaxWarning::DeprecatedListCatchAllPattern {
+                            location: SrcSpan { start, end: rsqb_e },
+                        })
+                }
 
                 Pattern::List {
                     location: SrcSpan { start, end },